Lagt in LOAD i currentState för att man ska kunna ladda en ny bana.
Har också fixat pausmenyn och tagit bort några varningar. :)

+++ Iris/Hud.hpp
@@ -15,7 +15,7 @@ private:
 	sf::Text goldText;
 	std::stringstream goldAmount;
 	sf::RectangleShape progressBar;
-	int progress;
+	float progress;
 	sf::Color progressColor;
 };
